chore: restructure ecmwf plugin into concrete sub-modules
