+++ backend/packages/fiab-plugin-ecmwf/src/fiab_plugin_ecmwf/blocks.py
@@ -8,13 +8,10 @@
 # nor does it submit to any jurisdiction.
 
 import logging
-import os
-import re
 
 import numpy as np
 from cascade.low.func import Either
 from earthkit.workflows.fluent import Action, Payload, from_source, merge
-from earthkit.workflows.nodetree import nodetree_dimensions, nodetree_new_dimension
 from fiab_core.fable import (
     ActionLookup,
     BlockConfigurationOption,
@@ -31,6 +28,14 @@ from pymetkit import ParamDB
 from qubed import Qube
 
 from .block_utils import (
+    _axis_value_strings,
+    _extract_dataset,
+    _is_empty_qube,
+    _param_id_to_param_key,
+    _param_key_to_param_id,
+    _parse_axis_value,
+)
+from .constants import (
     BASE_TIME,
     DIMENSION,
     DOMAIN,
@@ -46,32 +51,11 @@ from .block_utils import (
     SPLITBY,
     STEP,
     VALUES,
-    _axis_value_strings,
-    _extract_dataset,
-    _is_empty_qube,
-    _param_id_to_param_key,
-    _param_key_to_param_id,
-    _parse_axis_value,
 )
 from .datasets import load_datasets
 from .qubed_utils import axes, common_dimensions, contains, dimensions, expand, select
 
 logger = logging.getLogger(__name__)
-
-GRIB_ALIASES = {
-    "shortName": PARAM,
-    "paramId": PARAM,
-    "stepRange": STEP,
-    "level": LEVEL,
-}
-
-GRIB_MIME = "text/plain; fiab-format=gribdir"
-
-PLOT_FORMAT_TO_MIME: dict[str, str] = {
-    "png": "image/png",
-    "pdf": "application/pdf",
-    "svg": "image/svg+xml",
-}
 
 FORECAST_DATASETS = load_datasets()
 
@@ -187,51 +171,6 @@ class OperationalForecastSource(Source):
         return Either.ok(final_action)
 
 
-class ZarrSink(Sink):
-    title: str = "Zarr Sink"
-    description: str = "Write dataset to a zarr on the local filesystem"
-    configuration_options: dict[ConfigurationOptionId, BlockConfigurationOption] = {
-        PATH: BlockConfigurationOption(
-            title="Zarr Path",
-            description="Filesystem path where the zarr should be written",
-            value_type=StringType(),
-        )
-    }
-    inputs: list[str] = ["dataset"]
-
-    def validate(
-        self, block: BlockInstanceRich, inputs: dict[str, QubedOutput], restrictions: ConfigurationOptionRestriction
-    ) -> BlockInstanceOutput:
-        _extract_dataset(inputs, "dataset")
-        return RawOutput(type_fqn="bytes", mime_type="text/plain")
-
-    def compile(
-        self,
-        inputs: ActionLookup,
-        block: BlockInstanceRich,
-    ) -> Either[Action, Error]:  # type:ignore[invalid-argument] # semigroup
-        input_task = block.input_ids["dataset"]
-
-        temp_dim = nodetree_new_dimension(inputs[input_task].nodes)
-        action = (
-            inputs[input_task]
-            .flatten(new_dim=temp_dim, reset_coords=True)
-            .combine_branches(dim=temp_dim)
-            .concatenate(dim=temp_dim)
-            .map(
-                Payload(
-                    "fiab_plugin_ecmwf.runtime.sinks.write_zarr",
-                    kwargs={"path": block.config_as_str(PATH)},
-                    metadata={"environment": ["zarr"]},
-                )
-            )
-        )
-        return Either.ok(action)
-
-    def intersect(self, other: QubedOutput) -> bool:
-        return bool(dimensions(other))
-
-
 class Select(Transform):
     title: str = "Select"
     description: str = "Select values from one dimension of the input dataset"
@@ -312,173 +251,3 @@ class Select(Transform):
 
     def intersect(self, other: QubedOutput) -> bool:
         return bool(dimensions(other))
-
-
-class GribSink(Sink):
-    title: str = "GRIB Sink"
-    description: str = "Write dataset to a GRIB file on the local filesystem"
-    configuration_options: dict[ConfigurationOptionId, BlockConfigurationOption] = {
-        PATH: BlockConfigurationOption(
-            title="GRIB Path",
-            description="Filesystem path where the GRIB file should be written. Filename can contain template values from metadata in [] brackets.",
-            value_type=StringType(),
-        )
-    }
-    inputs: list[str] = ["dataset"]
-
-    def _find_template_values(cls, path: str) -> list[str]:
-        return re.findall(r"\[(.*?)\]", path)
-
-    def validate(
-        self, block: BlockInstanceRich, inputs: dict[str, QubedOutput], restrictions: ConfigurationOptionRestriction
-    ) -> BlockInstanceOutput:
-        _extract_dataset(inputs, "dataset")  # check format of input and existence of dataset
-        path = block.config_as_str(PATH)
-        dirname = os.path.dirname(path)
-        if len(self._find_template_values(dirname)) != 0:
-            raise ValueError("Invalid filepath: directory path can not contain template values")
-        return RawOutput(type_fqn="bytes", mime_type=GRIB_MIME)
-
-    def compile(
-        self,
-        inputs: ActionLookup,
-        block: BlockInstanceRich,
-    ) -> Either[Action, Error]:  # type:ignore[invalid-argument] # semigroup
-        input_task = block.input_ids["dataset"]
-        path_dims = self._find_template_values(block.config_as_str(PATH))
-        action_dims = nodetree_dimensions(inputs[input_task].nodes)
-        keep_dims = []
-        for dim in path_dims:
-            mapped_dim = GRIB_ALIASES.get(dim, dim)
-            if mapped_dim in action_dims and mapped_dim not in keep_dims:
-                keep_dims.append(mapped_dim)
-        temp_dim = nodetree_new_dimension(inputs[input_task].nodes)
-        action = inputs[input_task].flatten(new_dim=temp_dim, keep_dims=keep_dims, reset_coords=True).concatenate(dim=temp_dim)
-        try:
-            if PARAM in keep_dims:
-                action = action.combine_branches(dim=PARAM)
-            else:
-                action = action.combine_branches(dim=temp_dim).concatenate(dim=temp_dim)
-        except:
-            pass
-
-        action = action.map(
-            Payload(
-                "fiab_plugin_ecmwf.runtime.sinks.write_grib",
-                kwargs={"path": block.config_as_str(PATH)},
-            )
-        )
-        return Either.ok(action)
-
-    def intersect(self, other: QubedOutput) -> bool:
-        return bool(dimensions(other))
-
-
-class MapPlotSink(Sink):
-    title: str = "Map Plot"
-    description: str = "Render a geographic map using earthkit-plots"
-    configuration_options: dict[ConfigurationOptionId, BlockConfigurationOption] = {
-        PARAM: BlockConfigurationOption(
-            title="Parameters",
-            description="Parameters to select and plot (e.g. '2t', 'msl')",
-            value_type=ListType(ParameterType()),
-        ),
-        DOMAIN: BlockConfigurationOption(
-            title="Domain",
-            description="Area to display: auto (fit the data), global, a named region/country (select several to union), or a drawn bounding box",
-            value_type=GeoDomainType(),
-            default_value="global",
-        ),
-        FORMAT: BlockConfigurationOption(
-            title="Format",
-            description="Output image format",
-            value_type=ClosedEnumType(["png", "pdf", "svg"]),
-            default_value="png",
-        ),
-        GROUPBY: BlockConfigurationOption(
-            title="Group By",
-            description="Dimension to create subplots over",
-            value_type=ClosedEnumType(["valid_datetime", "step", "number", "none"]),
-            default_value="none",
-        ),
-        SPLITBY: BlockConfigurationOption(
-            title="Split By",
-            description="Dimensions to separate plots by",
-            value_type=ListType(StringType()),
-            default_value="step",
-        ),
-        # ConfigurationOptionId("style_schema"): BlockConfigurationOption(
-        #     title="Style Schema",
-        #     description="earthkit-plots schema identifier",
-        #     value_type=StringType(),
-        # ),
-    }
-    inputs: list[str] = ["dataset"]
-
-    def validate(
-        self, block: BlockInstanceRich, inputs: dict[str, QubedOutput], restrictions: ConfigurationOptionRestriction
-    ) -> BlockInstanceOutput:
-        input_dataset = _extract_dataset(inputs, "dataset")
-
-        input_axes = axes(input_dataset)
-        input_param_values = input_axes.get(PARAM, set())
-        param_values = [_param_id_to_param_key(x) for x in input_param_values]
-        if param_values:
-            restrictions[PARAM] = ListType(ClosedEnumType(sorted(param_values), subtype=ParameterType()))
-
-        common = common_dimensions(input_dataset).intersection({PARAM, STEP, ENSEMBLE, LEVEL})
-        splitby = [x for x in common if len(input_axes[x]) > 1]
-        restrictions[SPLITBY] = ListType(ClosedEnumType(sorted(splitby) + ["none"]))
-
-        params = block.config_as_list(PARAM, str, allow_empty=False)
-        splitby_value = block.config_as_list(SPLITBY, str, allow_empty=True)
-        fmt = block.config_as_str(FORMAT)
-
-        missing_params = [param for param in params if param not in param_values]
-        if missing_params:
-            raise ValueError(f"params {missing_params} are not in the input parameters: {param_values}")
-
-        if "none" in splitby_value and len(splitby_value) != 1:
-            raise ValueError("Invalid splitby value: if none is selected, no other dimensions can be present")
-
-        mime_type = PLOT_FORMAT_TO_MIME.get(fmt)
-        if mime_type is None:
-            raise ValueError(f"Unsupported output format: {fmt}")
-        return RawOutput(type_fqn="bytes", mime_type=mime_type)
-
-    def compile(
-        self,
-        inputs: ActionLookup,
-        block: BlockInstanceRich,
-    ) -> Either[Action, Error]:  # type:ignore[invalid-argument] # semigroup
-        input_task = block.input_ids["dataset"]
-        params = [_param_key_to_param_id(x) for x in block.config_as_list(PARAM, str, allow_empty=False)]
-        groupby = block.config_as_str(GROUPBY)
-        splitby = block.config_as_list(SPLITBY, str, allow_empty=True)
-        if "none" in splitby:
-            splitby = []
-
-        selected = (
-            inputs[input_task]
-            .combine_branches(dim=PARAM, force=True)
-            .select({PARAM: params if len(params) > 1 else params[0]})
-            .flatten(new_dim="temp_dim", keep_dims=splitby, reset_coords=True)
-            .concatenate(dim="temp_dim")
-        )
-
-        action = selected.map(
-            Payload(
-                "fiab_plugin_ecmwf.runtime.plots.map_plot",
-                kwargs={
-                    "domain": block.config_as_geodomain(DOMAIN).with_bbox_earthkitplots().value or None,
-                    "format": block.config_as_str(FORMAT),
-                    "groupby": None if groupby == "none" else groupby,
-                    # "style_schema": block.config_as_str("style_schema") or "inbuilt://fiab",
-                },
-                metadata={"environment": ["earthkit-plots<1.0.0", "earthkit-regrid<1.0.0", "matplotlib<3.11"]},
-            )
-        )
-        return Either.ok(action)
-
-    def intersect(self, other: QubedOutput) -> bool:
-        return contains(other, PARAM)
