+++ backend/packages/fiab-plugin-ecmwf/src/fiab_plugin_ecmwf/constants.py
@@ -1,0 +1,33 @@
+# (C) Copyright 2026- ECMWF.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+from fiab_core.fable import ConfigurationOptionId
+
+BASE_TIME = ConfigurationOptionId("base_time")
+CHECKPOINT = ConfigurationOptionId("checkpoint")
+COMPARISON = ConfigurationOptionId("comparison")
+DIMENSION = ConfigurationOptionId("dimension")
+DOMAIN = ConfigurationOptionId("domain")
+ENSEMBLE = ConfigurationOptionId("number")
+FORECAST = ConfigurationOptionId("forecast")
+FORMAT = ConfigurationOptionId("format")
+GROUPBY = ConfigurationOptionId("groupby")
+INPUT_SOURCE = ConfigurationOptionId("input_source")
+LEAD_TIME = ConfigurationOptionId("lead_time")
+LEVEL = ConfigurationOptionId("levelist")
+LEVTYPE = ConfigurationOptionId("levtype")
+PARAM = ConfigurationOptionId("param")
+PATH = ConfigurationOptionId("path")
+SOURCE = ConfigurationOptionId("source")
+SPLITBY = ConfigurationOptionId("splitby")
+STATISTIC = ConfigurationOptionId("statistic")
+STEP = ConfigurationOptionId("step")
+THRESHOLD = ConfigurationOptionId("threshold")
+TYPE = ConfigurationOptionId("type")
+VALUES = ConfigurationOptionId("values")
