+++ backend/packages/fiab-plugin-ecmwf/src/fiab_plugin_ecmwf/plotting/__init__.py
@@ -1,0 +1,13 @@
+# (C) Copyright 2026- ECMWF.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+
+from .blocks import MapPlotSink
+
+__all__ = ["MapPlotSink"]
