+++ backend/packages/fiab-plugin-ecmwf/tests/unit/test_blocks.py
@@ -41,7 +41,9 @@ from fiab_plugin_ecmwf.block_utils import (
     VALUES,
     _param_id_to_param_key,
 )
-from fiab_plugin_ecmwf.blocks import GribSink, MapPlotSink, OperationalForecastSource, Select, ZarrSink
+from fiab_plugin_ecmwf.blocks import OperationalForecastSource, Select
+from fiab_plugin_ecmwf.output import GribSink, ZarrSink
+from fiab_plugin_ecmwf.plotting import MapPlotSink
 from fiab_plugin_ecmwf.products.blocks import EnsembleStatistics
 from fiab_plugin_ecmwf.qubed_utils import axes, collapse, contains
